Fix a compilation warning with clang 3.0
In b1ad1a1d0266a20bb we introduced an implicit (but safe)
long-to-int shortening that clang didn't like.

Warning not in any released version of Tor.

+++ src/common/util.c
@@ -3347,7 +3347,7 @@ format_helper_exit_status(unsigned char child_state, int saved_errno,
   *cur++ = '\n';
   *cur++ = '\0';
 
-  res = cur - hex_errno - 1;
+  res = (int)(cur - hex_errno - 1);
 
   goto done;
